requirements_dev.txt
# Reading WiFi packages
pyshark
scapy

# Reading GPS csv_data
gpsd-py3

# Command Line
click

# Parsing data
numpy
